First of all we have another system function called "result". For now it just creates an internal variable for the return value. Later it will be replaced by an REAL return.
Started recursive macro calls support by adding a trace vector and doing an new evaluateFunction method.
The plan is to allow this only on evaluate macro parameters. As soon as reaching the macro call it will
move the ip to the original macro, run it with the given parameter, and then hopefully return.

Its wip. And its so far the biggest addition to the original bass.
Not sure its a good Idea. We could try to solve that by other tricks .. maybe.